+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ### Feat
 
+- **asset**: add type icon svg assets
+- **core**: add pokemon and some utility models
+- **ui**: add common ui widgets and pokemon colors
+- **routes**: add shell routes for pokemon, items and moves
+- **routes**: use go_router instead of bart for bottom bar rendering
+- **i18n**: update translations
 - **i18n**: update translation descriptions
 - **i18n**: initial localizations created
 
@@ -11,6 +17,8 @@
 
 ### Refactor
 
+- **routes**: remove remnants of previous router
+- rename environment file
 - **imports**: refactor and reorganize imports
 - **builds**: remove macos and ios builds
 - **appname**: change app name to YAPDex